+++ CHANGELOG.md
@@ -30,12 +30,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Size prop to Modal.
 - `aria-label`, `aria-labelledby`, and `aria-describedby` prop to Modal.
 - `FocusTrap` in modal.
-- `initialFocusRef` prop to Modal.
 
 ### Fixed
 
 - `centered` Modal prop when `false` did not changed.
 - ModalDialog documentation didn't informed the required props.
+- `autoFocus` prop in Input blocking navigation.
 
 ## [9.112.27] - 2020-03-26